Added button to open the diff tool
