+++ Source/GitHubTools/Private/Widgets/SGitHubToolsFileInfosRow.h
@@ -1,0 +1,33 @@
+#pragma once
+
+#include "SGitHubToolsPRInfos.h"
+
+#include <CoreMinimal.h>
+
+class SGitHubToolsFileInfosRow final : public STableRow< FGitHubToolsFileInfosTreeItemPtr >
+{
+public:
+    SLATE_BEGIN_ARGS( SGitHubToolsFileInfosRow )
+    {}
+
+    SLATE_ARGUMENT( FGitHubToolsFileInfosTreeItemPtr, TreeItem )
+    SLATE_ARGUMENT( FGithubToolsPullRequestInfosPtr, PRInfos )
+    SLATE_EVENT( FGitHubToolsPRInfosOnTreeItemStateChangedDelegate, OnTreeItemStateChanged )
+
+    SLATE_END_ARGS()
+
+    void Construct( const FArguments & arguments, const TSharedRef< STableViewBase > & owner_table_view );
+
+private:
+    FReply OnMarkAsViewedButtonClicked();
+    FReply OnOpenAssetButtonClicked();
+    FReply OnDiffAssetButtonClicked();
+    void OpenTreeItemAsset( FGitHubToolsFileInfosTreeItemPtr tree_item );
+
+    FGithubToolsPullRequestInfosPtr PRInfos;
+    FString PRId;
+    FGitHubToolsFileInfosTreeItemPtr TreeItem;
+    FGitHubToolsPRInfosOnTreeItemStateChangedDelegate OnTreeItemStateChanged;
+};
+
+#undef LOCTEXT_NAMESPACE
